refactor: remove rust_decimal dependency in favor of f64
SQLite REAL columns are IEEE 754 doubles, so the Decimal layer added
no precision benefit while requiring conversion wrappers at every DB
boundary. Replace all entity Decimal fields with f64, drop the
to_dec/from_dec helpers and the rust_decimal crate entirely.

+++ src-tauri/src/entities/fermentables.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "fermentables")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
@@ -12,8 +12,8 @@ pub struct Model {
     pub name: String,
     #[sea_orm(column_type = "Text")]
     pub r#type: String,
-    pub yield_pct: Decimal,
-    pub color_lovibond: Decimal,
+    pub yield_pct: f64,
+    pub color_lovibond: f64,
     #[sea_orm(column_type = "Text", nullable)]
     pub origin: Option<String>,
     #[sea_orm(column_type = "Text", nullable)]
@@ -21,13 +21,13 @@ pub struct Model {
     #[sea_orm(column_type = "Text", nullable)]
     pub notes: Option<String>,
     pub add_after_boil: Option<i32>,
-    pub coarse_fine_diff_pct: Option<Decimal>,
-    pub moisture_pct: Option<Decimal>,
-    pub diastatic_power_lintner: Option<Decimal>,
-    pub protein_pct: Option<Decimal>,
-    pub max_in_batch_pct: Option<Decimal>,
+    pub coarse_fine_diff_pct: Option<f64>,
+    pub moisture_pct: Option<f64>,
+    pub diastatic_power_lintner: Option<f64>,
+    pub protein_pct: Option<f64>,
+    pub max_in_batch_pct: Option<f64>,
     pub recommend_mash: Option<i32>,
-    pub ibu_gal_per_lb: Option<Decimal>,
+    pub ibu_gal_per_lb: Option<f64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
