+++ src-tauri/src/entities/miscs.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "miscs")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
@@ -14,7 +14,7 @@ pub struct Model {
     pub r#type: String,
     #[sea_orm(column_type = "Text")]
     pub r#use: String,
-    pub time_min: Decimal,
+    pub time_min: f64,
     #[sea_orm(column_type = "Text", nullable)]
     pub notes: Option<String>,
     #[sea_orm(column_type = "Text", nullable)]
