+++ src-tauri/src/entities/recipe_addition_miscs.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "recipe_addition_miscs")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
@@ -18,9 +18,9 @@ pub struct Model {
     pub r#type: String,
     #[sea_orm(column_type = "Text")]
     pub r#use: String,
-    pub amount: Decimal,
+    pub amount: f64,
     pub amount_is_weight: Option<i32>,
-    pub time_min: Decimal,
+    pub time_min: f64,
     pub addition_order: i32,
 }
