+++ src-tauri/src/entities/recipe_addition_waters.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "recipe_addition_waters")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
@@ -14,7 +14,7 @@ pub struct Model {
     pub water_id: Option<String>,
     #[sea_orm(column_type = "Text")]
     pub name: String,
-    pub amount_l: Decimal,
+    pub amount_l: f64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
